+++ src/app/components/leavetype/leavetype.component.ts
@@ -1,9 +1,10 @@
 import { Component,OnInit } from '@angular/core';
-import { MessageService } from 'primeng/api';
+import { ConfirmEventType,ConfirmationService, MessageService } from 'primeng/api';
 import { map } from 'rxjs';
 import { LeaveType } from 'src/app/models/leavetype';
 import { leavetypeService } from 'src/app/services/leavetype.service';
 declare var dataTableInit:any; 
+
 @Component({
   selector: 'app-leavetype',
   templateUrl: './leavetype.component.html',
@@ -17,10 +18,13 @@ export class LeaveTypeComponent  implements OnInit{
     display = false;
     sibebarHeader = "Add LeaveType";
     selectedKey=''
+
     constructor(
       public leavetypeservice: leavetypeService,
       private messageService: MessageService,
+      private confirmationService:ConfirmationService
     ){ }
+
       ngOnInit():void {
         this.cols = [
           { field: 'type', header: 'type', customExportHeader: 'TYPE' },
@@ -45,7 +49,7 @@ export class LeaveTypeComponent  implements OnInit{
           }
           else{
             this.leavetypeservice.create(this.leavetype).then(() => {
-              this.messageService.add({ severity: 'success', summary: 'Success', detail:'Leavetype is edited succcessfully' });
+              this.messageService.add({ severity: 'success', summary: 'Success', detail:'Leavetype is created succcessfully' });
               this.display = false;
               this.leavetype = {Type:''}
             });
@@ -63,6 +67,27 @@ export class LeaveTypeComponent  implements OnInit{
         
           }
           delete(key:any){
-            this.leavetypeservice.delete(key);
+            this.confirmationService.confirm({
+              message: 'Do you want to delete this record?',
+              header: 'Delete Confirmation',
+              icon: 'pi pi-info-circle',
+              accept: () => {
+                this.leavetypeservice.delete(key);
+                this.messageService.add({ severity: 'info', summary: 'Confirmed', detail: 'Record deleted' });
+              },
+              reject: (type) => {
+                  switch (type) {
+                      case ConfirmEventType.REJECT:
+                          this.messageService.add({ severity: 'error', summary: 'Rejected', detail: 'You have rejected' });
+                          break;
+                      case ConfirmEventType.CANCEL:
+                          this.messageService.add({ severity: 'warn', summary: 'Cancelled', detail: 'You have cancelled' });
+                          break;
+                  }
+                }
+              }); 
+          }
+          sidenavClosed(){
+            this.leavetype = {Type:''}
           }
 }
